Protocol Specification Started
I have decided to restructure the way that the documentation in this
project is managed. The doc/ folder will now be reserved for ideas and
a place to store useful files relating to this project. The
specification/ folder contains a formal definition of the protocol and
a set of tools to build a portable protocol specification. Both of
these folders will grow and change as the protocol is further defined.